fixed: vagrant build

+++ setup/bootstrap.sh
@@ -20,31 +20,33 @@ echo 'updating enketo app to latest version'
 apt-get install -y git
 cd $ENKETO_EXPRESS_REPO_DIR
 # The next line should be commented out during development or any case where the repository was cloned via the ssh URL (i.e. git@github.com:kobotoolbox/enketo-express.git) as opposed to the HTTPS.
-#[ $ENKETO_EXPRESS_UPDATE_REPO = "true" ] && git pull
-#git submodule update --init --recursive
+[ $ENKETO_EXPRESS_UPDATE_REPO = "true" ] && git pull
+git submodule update --init --recursive
 
 # further redis setup with persistence, security, logging, multiple instances, priming 
-stop redis-server
+#stop redis-server
 echo 'copying enketo redis conf...'
-mv /etc/redis/redis.conf /etc/redis/redis-origin.conf
-cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/conf/redis-enketo-main.conf /etc/redis/
-cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/conf/redis-enketo-cache.conf /etc/redis/
-chown redis:redis /var/lib/redis/
-echo 'copying enketo redis-server configs...'
-mv /etc/init/redis-server.conf /etc/init/redis-server.conf.disabled
-cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/init/redis-server-enketo-main.conf /etc/init/
-cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/init/redis-server-enketo-cache.conf /etc/init/
-if [ -f "/var/lib/redis/redis.rdb" ]; then
-	rm /var/lib/redis/redis.rdb
+if [ -f "/etc/redis/redis.conf" ]; then
+    mv /etc/redis/redis.conf /etc/redis/redis-origin.conf
+    cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/conf/redis-enketo-main.conf /etc/redis/
+    cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/conf/redis-enketo-cache.conf /etc/redis/
+    chown redis:redis /var/lib/redis/
+    echo 'copying enketo redis-server configs...'
+    mv /etc/init/redis-server.conf /etc/init/redis-server.conf.disabled
+    cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/init/redis-server-enketo-main.conf /etc/init/
+    cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/init/redis-server-enketo-cache.conf /etc/init/
+    if [ -f "/var/lib/redis/redis.rdb" ]; then
+	   rm /var/lib/redis/redis.rdb
+    fi
+    echo 'copying enketo default redis db...'
+    cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/enketo-main.rdb /var/lib/redis/
+    chown redis:redis /var/lib/redis/enketo-main.rdb
+    chmod 660 /var/lib/redis/enketo-main.rdb
 fi
-echo 'copying enketo default redis db...'
-cp -f $ENKETO_EXPRESS_REPO_DIR/setup/redis/enketo-main.rdb /var/lib/redis/
-chown redis:redis /var/lib/redis/enketo-main.rdb
-chmod 660 /var/lib/redis/enketo-main.rdb
 echo 'starting first enketo redis instance...'
-service redis-server-enketo-main start
+service redis-server-enketo-main restart
 echo 'starting second enketo redis instance...'
-service redis-server-enketo-cache start
+service redis-server-enketo-cache restart
 
 # install XML prerequisites for node_xslt
 echo 'installing libxml2 and libxslt'
@@ -63,18 +65,19 @@ else
     apt-get update
     apt-get install -y nodejs
 fi
-npm install -g grunt-cli nodemon mocha bower
+npm -g install grunt-cli nodemon mocha bower
+npm -g install npm@next
 # remove node_modules if exists because npm builds can be system-specific
 if [ -d "$ENKETO_EXPRESS_REPO_DIR/node_modules" ]; then
 	rm -R $ENKETO_EXPRESS_REPO_DIR/node_modules
 fi
-npm install 
+npm install
 bower install --allow-root
 
 # create a configuration file by copying the default unless it already exists
-echo 'copying default configuration unless config.json already exists'
-if [ -f "$ENKETO_EXPRESS_REPO_DIR/config/config.json"]; then
-    cp config/default-config.json config/config.json
+echo 'copying custom configuration unless config.json already exists'
+if [ ! -f "$ENKETO_EXPRESS_REPO_DIR/config/config.json" ]; then
+    cp setup/config/config.json config/config.json
 fi
 
 # build js and css
